+++ src/jflow/db/instdata/tests/api.py
@@ -1,0 +1,11 @@
+import datetime
+
+from django.test import TestCase
+from django.conf import settings
+
+
+class CalendarViewTest(TestCase):
+    fixtures = ["vendor.json"]
+        
+    def testVersion(self):
+        pass