+++ CHANGELOG.rst
@@ -1,8 +1,16 @@
-jflow 0.2
-============
-first unofficial release
 
-jflow 0.3-alpha
+jflow 0.4-pre-alpha
 ===================
+ * Refactoring `instdata` and `trade` models. Not backward compatible
+
+
+jflow 0.3   2010-03-15
+==========================
+ * First tagged release.
  * Added small management utility for cleaning stale instruments (python manage.py cleaninst)
  * Added TabularInline BondIssuer in BondClass admin
+
+
+jflow 0.2
+==========================
+first unofficial release